requirements.txt
pandas
numpy
scikit-learn
matplotlib
kagglehub
seaborn
jupyter
ipykernel
azureml-core
azureml-defaults